Change transition matrix G
Changed G from a conditional transition matrix to a "stationary" matrix (the initial distribution). My understanding was that we could then use G as H, but I am not 100% sure. I only changed the definition of G here, so the code for solving for H will not work because I didn't change that.